+++ src/main/java/ru/practicum/shareit/request/service/RequestService.java
@@ -3,5 +3,5 @@ package ru.practicum.shareit.request.service;
 import ru.practicum.shareit.request.model.ItemRequest;
 
 public interface RequestService {
-    Integer addRequest(ItemRequest request);
+    Long addRequest(ItemRequest request);
 }
