+++ src/main/java/ru/practicum/shareit/request/storage/RequestStore.java
@@ -3,5 +3,5 @@ package ru.practicum.shareit.request.storage;
 import ru.practicum.shareit.request.dto.ItemRequestDTO;
 
 public interface RequestStore {
-    Integer addRequest(ItemRequestDTO.NewReturnItemRequestDTO requestDTO);
+    Long addRequest(ItemRequestDTO.NewReturnItemRequestDTO requestDTO);
 }
